Split legal banners into separate files
